requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
deprecation==2.1.0
exceptiongroup==1.2.0
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
numpy==1.26.4
packaging==24.0
pandas==2.2.1
postgrest==0.16.1
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
storage3==0.7.1
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websockets==11.0.3
